Fix various bugs in BB8
